backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infra-autopilot"
version = "1.0.0"
description = "Infrastructure prerequisite validation system"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # LLM + Extraction pipeline
    "openai>=1.35.0",
    "instructor>=1.3.0",
    "langgraph>=0.1.0",

    # Document parsers
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.2",
    "docling>=2.0.0",

    # Azure SDK
    "azure-identity>=1.17.0",
    "azure-mgmt-compute>=31.0.0",
    "azure-mgmt-containerservice>=30.0.0",
    "azure-mgmt-storage>=21.1.0",
    "azure-mgmt-rdbms>=10.1.0",
    "azure-mgmt-network>=25.3.0",
    "azure-mgmt-authorization>=4.0.0",
    "azure-mgmt-msi>=7.0.0",
    "azure-mgmt-resource>=23.1.0",
    "azure-mgmt-keyvault>=10.3.0",
    "azure-keyvault-certificates>=4.8.0",

    # Logging
    "structlog>=24.2.0",

    # Utilities
    "packaging>=24.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
